- Updated resource manager to use uint_fast64_t instead.

+++ src/game/src/managers/resource/resource_manager.h
@@ -15,15 +15,15 @@ enum RESOURCE : uint_fast8_t
 class resource_manager
 {
 private:
-    std::unordered_map<RESOURCE, int> resources;
+    std::unordered_map<RESOURCE, uint_fast64_t> resources;
 
 public:
     resource_manager();
-    resource_manager(std::unordered_map<RESOURCE, int> initialising_resources);
+    resource_manager(std::unordered_map<RESOURCE, uint_fast64_t> initialising_resources);
 
-    int get_resource_amount(const RESOURCE resource) const;
-    bool has_enough_resource(const RESOURCE resource, int amount) const;
+    uint_fast64_t get_resource_amount(const RESOURCE resource) const;
+    bool has_enough_resource(const RESOURCE resource, uint_fast64_t amount) const;
 
-    void add_resource(const RESOURCE resource, int amount);
-    bool remove_resource(const RESOURCE resource, int amount);
+    void add_resource(const RESOURCE resource, uint_fast64_t amount);
+    bool remove_resource(const RESOURCE resource, uint_fast64_t amount);
 };
